Modificamos dato Monto a tipo float

+++ Models/Metas.cs
@@ -11,6 +11,6 @@ namespace Parcial1_Ap1_VladimirGuzman.Models
         public string? Descripcion { get; set; }
         [Required(ErrorMessage ="Desbes ingresar el monto")]
         [Range(1, int.MaxValue, ErrorMessage ="No puede ser un monto Menor que 0")]
-        public int Monto { get; set; }
+        public float Monto { get; set; }
     }
 }
